Made the bot standalone. Lots of (breaking) changes and improvements.
The bot started out as a project for one Discord server and was intended to be a part of other bots (as a plugin).
Since then, its status has rose and it's been used standalone. I also learned some Docker usage in meantime.
That's why, in order to evolve and make this bot even better, I decided to make a lot of (breaking) changes.

Changelog:

- Changed the environment variable names
Every environment variable used by the bot now has the "MAMI_" prefix.

- Changed the DB table name from just "configs" to "mami_server_configs"
- Changed the default SQLite3 database filename from `spoilerbot_configs.db` to `mami_server_configs.db`

This change was made so that the bot can connect to an external database with lots of tables.
The filename is consistent with that.

- Moved SpoilerBotEncoder to another file; renamed it to SpoilerEncoder
- Separated SpoilerEncoder spec tests from the bot spec tests

This library actually encodes/decodes spoilers, not spoiler bots ;)
Keeping this separate from the actual bot makes sense.

- Fixed code formatting and style

It should be more consise now.

- Updated README
What else could you expect from an update this big?

- Updated the commands
- Made it possible to change the prefix
- Changed required permissions

The command syntax has been changed.
The default prefix is still "!", but can be changed by setting an environment variable.
You no longer have to be an administrator to be able to use the bot's commands - a moderator (the "Manage Channels" permission) will do.