More customization. Wrote the "help" command.
- Made it possible to change the command name
The users are no longer locked to the default "!mami" command; they can set even something as crazy as "~spoil" to invoke the bot.

- Created the "help" command
Disabled the discordrb's built in "help" functionality in favor of my own. Put the whole bot's usage in one nice embed.

- Modified the "display_config" a little
It now uses an embed instead of a regular message.

- Minor refactoring